Brenna Chen - Sprint #3 Pull Request (#10)
* hello

* Added video loading screen

* Delete .env

* Questions Screen (#8)

Squashing to remove .env file.

* pull request edits

* pull request edits

* questions

* changed MainTabNavigator
