+++ package.json
@@ -51,6 +51,7 @@
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react-native-animated-ellipsis": "^2.0.0",
     "react-native-gesture-handler": "~1.3.0",
     "react-native-screens": "^1.0.0-alpha.23",
     "react-native-video": "^5.0.2",
