+++ src/screens/Home/RenderingVideoScreen/styles.tsx
@@ -1,0 +1,41 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fff',
+    color: '#000',
+  },
+  title: {
+    fontSize: 40,
+  },
+  main: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  ellipses: {
+    color: '#E5186E',
+    fontSize: 200,
+    letterSpacing: -50,
+  },
+  text: {
+    fontFamily: 'montserrat-regular', 
+    fontStyle: 'normal',
+    fontWeight: 'bold',
+    fontSize: 30,
+    lineHeight: 37,
+    textAlign: 'center',
+    color: '#E5186E',
+    height: 114,
+    width: 307,
+    marginBottom: '45%',
+  },
+})
+
+export default styles
